+++ test/controllers/pages_controller_test.rb
@@ -13,4 +13,23 @@ class PagesControllerTest < ActionDispatch::IntegrationTest
     get root_url
     assert_response :success
   end
+
+  test "home invites a signed-out visitor to sign in" do
+    get root_url
+
+    assert_select "a", text: "Sign in"
+    assert_select "form[action=?]", session_path, count: 0
+  end
+
+  test "home greets a signed-in visitor and offers sign out" do
+    sign_in :kerrick
+
+    get root_url
+
+    assert_response :success
+    assert_match users(:kerrick).name, response.body
+    assert_select "form[action=?]", session_path do
+      assert_select "button", text: "Sign out"
+    end
+  end
 end